fix(QuestionService) : Added validation errors

+++ src/Services/QuestionService.cs
@@ -47,10 +47,30 @@ namespace Qubiz.QuizEngine.Services
 			}
 		}
 
-		public async Task UpdateQuestionAsync(QuestionDetail question)
+		public async Task<ValidationError[]> UpdateQuestionAsync(QuestionDetail question)
 		{
 			using (IUnitOfWork unitOfWork = unitOfWorkFactory.Create())
 			{
+				if (question.QuestionText.Equals(String.Empty))
+					return new ValidationError[1] { new ValidationError { Message = "Question text can't be empty" } };
+
+				if (question.Options.Length < 3)
+					return new ValidationError[1] { new ValidationError { Message = "Question must have at last 3 options !" } };
+
+				if (question.SectionID == Guid.Empty)
+					return new ValidationError[1] { new ValidationError { Message = "Question must have a section assigned !" } };
+
+				foreach (OptionDefinition option in question.Options)
+				{
+					if (option.Answer.Equals(String.Empty))
+					{
+						return new ValidationError[1] { new ValidationError { Message = "Option answer can't be empty !" } };
+					}
+				}
+
+				if (!question.Options.Any(x => x.IsCorrectAnswer == true))
+					return new ValidationError[1] { new ValidationError { Message = "Options must have at last 1 correct answer !" } };
+
 				await unitOfWork.QuestionRepository.UpdateQuestionAsync(question.DeepCopyTo<Database.Models.QuestionDefinition>());
 				unitOfWork.OptionRepository.DeleteOptionsAsync((await unitOfWork.OptionRepository.GetOptionsByQuestionIDAsync(question.ID)).ToArray());
 				unitOfWork.OptionRepository.AddOptionsAsync(question.Options.Select(o => new Database.Models.OptionDefinition
@@ -62,13 +82,34 @@ namespace Qubiz.QuizEngine.Services
 					QuestionID = o.QuestionID
 				}).ToArray());
 				await unitOfWork.SaveAsync();
+
+				return new ValidationError[0];
 			}
 		}
 
-		public async Task AddQuestionAsync(QuestionDetail question)
+		public async Task<ValidationError[]> AddQuestionAsync(QuestionDetail question)
 		{
 			using (IUnitOfWork unitOfWork = unitOfWorkFactory.Create())
 			{
+				if (question.QuestionText.Equals(String.Empty))
+					return new ValidationError[1] { new ValidationError { Message = "Question text can't be empty" } };
+
+				if (question.Options.Length < 3)
+					return new ValidationError[1] { new ValidationError { Message = "Question must have at last 3 options !" } };
+
+				if (question.SectionID == Guid.Empty)
+					return new ValidationError[1] { new ValidationError { Message = "Question must have a section assigned !" } };
+
+				foreach (OptionDefinition option in question.Options)
+				{
+					if (option.Answer.Equals(String.Empty))
+						return new ValidationError[1] { new ValidationError { Message = "Option answers can't be empty !" } };
+
+				}
+
+				if (!question.Options.Any(x => x.IsCorrectAnswer == true))
+					return new ValidationError[1] { new ValidationError { Message = "Options must have at last 1 correct answer !" } };
+
 				await unitOfWork.QuestionRepository.AddQuestionAsync(question.DeepCopyTo<Database.Models.QuestionDefinition>());
 				unitOfWork.OptionRepository.AddOptionsAsync(question.Options.Select(o => new Database.Models.OptionDefinition
 				{
@@ -79,6 +120,7 @@ namespace Qubiz.QuizEngine.Services
 					QuestionID = o.QuestionID
 				}).ToArray());
 				await unitOfWork.SaveAsync();
+				return new ValidationError[0];
 			}
 		}
